refactor: change player and match IDs from Long to int

+++ src/main/java/murlov/tennis_scoreboard/model/PlayerScore.java
@@ -6,14 +6,14 @@ import lombok.Setter;
 @Getter
 @Setter
 public class PlayerScore {
-    private final Long id;
+    private final int id;
     private final String name;
     private GamePoints points;
     private int games;
     private int sets;
     private Integer tieBreakPoints;
 
-    public PlayerScore(Long id, String name) {
+    public PlayerScore(int id, String name) {
         this.id = id;
         points = GamePoints.ZERO;
         this.name = name;
